+++ sheerwater_benchmarking/forecasts/__init__.py
@@ -1,5 +1,5 @@
 """Forecasting models for the Sheerwater benchmarking project."""
-from .salient import salient, salient_daily_gap
+from .salient import salient
 from .abc import perpp
 from .ecmwf_er import ecmwf_ifs_er, ecmwf_ifs_er_debiased
