+++ sheerwater_benchmarking/reanalysis/era5.py
@@ -102,7 +102,7 @@ def era5_daily(start_time, end_time, variable, grid="global1_5"):
                }
            },
            auto_rechunk=False)
-def era5_daily_regrid(start_time, end_time, variable, method="conservative", grid="global0_25"):
+def era5_daily_regrid(start_time, end_time, variable, grid="global0_25"):
     """ERA5 daily reanalysis with regridding."""
     ds = era5_daily(start_time, end_time, variable, grid='global0_25')
     ds = ds.sortby('lat')  # TODO: remove if we fix the era5 daily caches
