+++ sheerwater_benchmarking/tasks/__init__.py
@@ -1,0 +1,5 @@
+"""Tasks models for the Sheerwater benchmarking project."""
+from .spw import rainy_season_onset_truth, rainy_season_onset_forecast
+
+# Use __all__ to define what is part of the public API.
+__all__ = ["rainy_season_onset_truth", "rainy_season_onset_forecast"]
